Divide package generator's tests into more files
